// Task Management Application/src/models/registerformeinterface/registerforminterface.tsx
export interface FormInterface{
    firstName: string
    lastName: string
    email:string
    password?: number | string
    confirmPassword?: number | string
    type?: string | boolean
    name: string
    placeholder: string 
    value: number
    text: string
    register: object
    errormessages:string
}